feat: Add Dockerfile for ARM-based Macs
Although gevent recently began to support arm64 binary wheels, they
are only supported on manylinux2014-compatible systems - which
Alpine is not. (Per: https://www.gevent.org/install.html) As
specified in the gevent install docs, when a binary wheel is not
available, pip will fall back on building from source.

Add a Dockerfile for ARM-based Macs that includes the dependencies
for building gevent.

Credit to @st0w for the change.

Closes #74